edit Document domain nad Repo

+++ src/main/java/com/example/es/repository/DocumentRepository.java
@@ -3,7 +3,9 @@ package com.example.es.repository;
 import com.example.es.domain.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentRepository extends JpaRepository<Document, Integer> {
+
+
+public interface DocumentRepository extends JpaRepository<Document, Long> {
 
 
 }
